backend/requirements.txt
django
channels
channels-redis
daphne
solana
solders
base58
djangorestframework
psycopg2-binary 
python-dotenv
django-cors-headers
django_redis
requests
aiohttp
adrf